requirements.txt
acme==1.29.0
bleach==5.0.1
certbot==1.29.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
commonmark==0.9.1
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==37.0.4
distro==1.7.0
docutils==0.19
googleapis-common-protos==1.56.4
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
josepy==1.13.0
keyring==23.7.0
parsedatetime==2.6
pkginfo==1.8.3
protobuf==3.20.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
pytz==2022.1
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
twine==4.0.1
urllib3==1.26.11
webencodings==0.5.1
yandexcloud==0.175.0
zipp==3.8.1
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
